stop workers as early as possible during application shutdown sequence
The disposable bean destroy method is called pretty late during app shutdown. The impact
is that workers are not properly stopped in certain cases. We can stop all workers in a
more graceful way by using `ApplicationListener<ContextClosedEvent>` interface combined
with `Ordered` with a high precedence. Like this, the `stopAllWorkers` method is called
when the grails environment is still fully there.